requirements.txt
streamlit>=1.30.0
requests==2.31.0
google-cloud-bigquery==3.11.4
google-auth==2.23.0
pandas>=2.0.0
plotly>=5.18.0
Pillow>=10.0.0
db-dtypes>=1.1.0
streamlit-authenticator>=0.3.3
pyyaml
bcrypt>=4.0.0
openpyxl
streamlit-autorefresh
extra-streamlit-components>=0.1.71